requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
requests==2.31.0
cryptography==44.0.0
pyOpenSSL==25.0.0
